monolithic pipeline refactor
