+++ src/streamdiffusion/controlnet/controlnet_sdxlturbo_pipeline.py
@@ -4,16 +4,14 @@ from PIL import Image
 import numpy as np
 from pathlib import Path
 
-from diffusers.models import ControlNetModel
-from diffusers.utils import load_image
 from diffusers import StableDiffusionXLPipeline, AutoencoderKL, AutoencoderTiny
 
 from ..pipeline import StreamDiffusion
 from .config import ControlNetConfig, StreamDiffusionControlNetConfig
-from .preprocessors import get_preprocessor
+from .base_controlnet_pipeline import BaseControlNetPipeline
 
 
-class SDXLTurboControlNetPipeline:
+class SDXLTurboControlNetPipeline(BaseControlNetPipeline):
     """
     SDXL Turbo ControlNet pipeline using StreamDiffusion
     
@@ -21,170 +19,10 @@ class SDXLTurboControlNetPipeline:
     using t_index_list for efficient real-time generation at high resolution.
     """
     
-    def __init__(self, 
-                 stream_diffusion: StreamDiffusion,
-                 device: str = "cuda",
-                 dtype: torch.dtype = torch.float16):
-        """
-        Initialize SDXL Turbo ControlNet pipeline
-        
-        Args:
-            stream_diffusion: Base StreamDiffusion instance
-            device: Device to run ControlNets on
-            dtype: Data type for ControlNet models
-        """
-        self.stream = stream_diffusion
-        self.device = device
-        self.dtype = dtype
-        
-        # ControlNet storage
-        self.controlnets: List[ControlNetModel] = []
-        self.controlnet_images: List[Optional[torch.Tensor]] = []
-        self.controlnet_scales: List[float] = []
-        self.preprocessors: List[Optional[Any]] = []
-        
-        # Store original unet_step method for patching
-        self._original_unet_step = None
-        self._is_patched = False
-        
-        # Optimization: Cache transforms and reusable tensors
-        import torchvision.transforms as transforms
-        self._cached_transform = transforms.ToTensor()
-        self._temp_tensor_cache = {}
-        
-        # Cache for preprocessed images to eliminate redundant processing
-        self._preprocessed_cache = {}
-        self._last_input_frame = None
-    
-    def add_controlnet(self, 
-                      controlnet_config: ControlNetConfig,
-                      control_image: Optional[Union[str, Image.Image, np.ndarray, torch.Tensor]] = None) -> int:
-        """
-        Add a ControlNet to the pipeline
-        
-        Args:
-            controlnet_config: ControlNet configuration
-            control_image: Control image (optional)
-            
-        Returns:
-            Index of the added ControlNet
-        """
-        if not controlnet_config.enabled:
-            print(f"ControlNet {controlnet_config.model_id} is disabled, skipping")
-            return -1
-        
-        # Load ControlNet model
-        print(f"Loading SDXL Turbo ControlNet: {controlnet_config.model_id}")
-        controlnet = self._load_controlnet_model(controlnet_config.model_id)
-        
-        # Load preprocessor if specified
-        preprocessor = None
-        if controlnet_config.preprocessor:
-            preprocessor = get_preprocessor(controlnet_config.preprocessor)
-            # Set preprocessor parameters including device and dtype
-            preprocessor_params = {
-                'device': self.device,
-                'dtype': self.dtype
-            }
-            if controlnet_config.preprocessor_params:
-                preprocessor_params.update(controlnet_config.preprocessor_params)
-            preprocessor.params.update(preprocessor_params)
-            # Update device and dtype directly
-            if hasattr(preprocessor, 'device'):
-                preprocessor.device = self.device
-            if hasattr(preprocessor, 'dtype'):
-                preprocessor.dtype = self.dtype
-        
-        # Process control image if provided
-        processed_image = None
-        if control_image is not None:
-            processed_image = self._prepare_control_image(control_image, preprocessor)
-        elif controlnet_config.control_image_path:
-            # Load from configured path
-            control_image = load_image(controlnet_config.control_image_path)
-            processed_image = self._prepare_control_image(control_image, preprocessor)
-        
-        # Add to collections
-        self.controlnets.append(controlnet)
-        self.controlnet_images.append(processed_image)
-        self.controlnet_scales.append(controlnet_config.conditioning_scale)
-        self.preprocessors.append(preprocessor)
-        
-        # Patch the StreamDiffusion pipeline if this is the first ControlNet
-        if len(self.controlnets) == 1:
-            self._patch_stream_diffusion()
-        
-        print(f"Added SDXL Turbo ControlNet {len(self.controlnets) - 1}: {controlnet_config.model_id}")
-        return len(self.controlnets) - 1
-    
-    def update_control_image_efficient(self, control_image: Union[str, Image.Image, np.ndarray, torch.Tensor]) -> None:
-        """
-        Efficiently update all ControlNets with cache-aware preprocessing
-        
-        Args:
-            control_image: New control image to apply to all ControlNets
-        """
-        # Check if we need to reprocess
-        if self._last_input_frame is not None:
-            if isinstance(control_image, (torch.Tensor, np.ndarray)) and isinstance(self._last_input_frame, type(control_image)):
-                if control_image is self._last_input_frame:
-                    return  # Same object, use cached results
-            elif isinstance(control_image, Image.Image) and isinstance(self._last_input_frame, Image.Image):
-                if control_image is self._last_input_frame:
-                    return  # Same object, use cached results
-        
-        self._last_input_frame = control_image
-        self._preprocessed_cache.clear()
-        
-        # Convert to tensor early for GPU processing
-        try:
-            control_tensor = self._convert_to_tensor_early(control_image)
-            use_tensor_processing = True
-        except Exception as e:
-            use_tensor_processing = False
-            print(f"⚠️  Tensor conversion failed: {e}, falling back to PIL processing")
-        
-        # Group ControlNets by preprocessor type to avoid duplicate processing
-        preprocessor_groups = {}
-        for i in range(len(self.controlnets)):
-            if self.controlnet_scales[i] > 0:  # Only process active ones
-                preprocessor = self.preprocessors[i]
-                if preprocessor is not None:
-                    preprocessor_type = type(preprocessor).__name__
-                    if preprocessor_type not in preprocessor_groups:
-                        preprocessor_groups[preprocessor_type] = {
-                            'preprocessor': preprocessor,
-                            'indices': []
-                        }
-                    preprocessor_groups[preprocessor_type]['indices'].append(i)
-        
-        # Process once per preprocessor type
-        for preprocessor_type, group in preprocessor_groups.items():
-            preprocessor = group['preprocessor']
-            
-            # Use tensor processing if available and input is tensor
-            has_tensor_processing = hasattr(preprocessor, 'process_tensor')
-            using_tensor = use_tensor_processing and has_tensor_processing
-            
-            if using_tensor:
-                processed_image = self._prepare_control_image(control_tensor, preprocessor)
-            else:
-                processed_image = self._prepare_control_image(control_image, preprocessor)
-            
-            # Cache the result
-            self._preprocessed_cache[preprocessor_type] = processed_image
-            
-            # Apply to all ControlNets using this preprocessor
-            for index in group['indices']:
-                self.controlnet_images[index] = processed_image
-    
-    def update_controlnet_scale(self, index: int, scale: float) -> None:
-        """Update the conditioning scale for a specific ControlNet"""
-        if 0 <= index < len(self.controlnets):
-            self.controlnet_scales[index] = scale
-        else:
-            raise IndexError(f"ControlNet index {index} out of range")
-    
+    def _get_model_type(self) -> str:
+        """Return the model type string for logging purposes"""
+        return "SDXL Turbo"
+
     @property
     def controlnet_configs(self) -> List[Dict[str, Any]]:
         """
@@ -201,442 +39,112 @@ class SDXLTurboControlNetPipeline:
                 'preprocessor': type(self.preprocessors[i]).__name__ if i < len(self.preprocessors) and self.preprocessors[i] is not None else None
             })
         return configs
-    
-    def get_last_processed_image(self, index: int) -> Optional[Image.Image]:
-        """Get the last processed control image for display purposes"""
-        if not (0 <= index < len(self.controlnets)):
-            return None
+
+    def _post_process_control_image(self, control_tensor: torch.Tensor) -> torch.Tensor:
+        """
+        Post-process control image tensor with SDXL-specific handling
         
-        preprocessor = self.preprocessors[index]
-        if preprocessor is None:
-            return None
+        Args:
+            control_tensor: Processed control image tensor
             
-        preprocessor_type = type(preprocessor).__name__
-        cached_result = self._preprocessed_cache.get(preprocessor_type)
-        
-        if cached_result is None:
-            return None
-        
-        # Handle tensor results from GPU processing
-        if isinstance(cached_result, torch.Tensor):
-            if hasattr(preprocessor, 'tensor_to_pil'):
-                return preprocessor.tensor_to_pil(cached_result)
-            else:
-                return self._tensor_to_pil_fallback(cached_result)
-        
-        return cached_result
-    
-    def _load_controlnet_model(self, model_id: str) -> ControlNetModel:
-        """Load a ControlNet model from HuggingFace or local path"""
-        try:
-            if Path(model_id).exists():
-                controlnet = ControlNetModel.from_pretrained(
-                    model_id,
-                    torch_dtype=self.dtype,
-                    local_files_only=True
-                )
-            else:
-                if "/" in model_id and model_id.count("/") > 1:
-                    # Handle subfolder case
-                    parts = model_id.split("/")
-                    repo_id = "/".join(parts[:2])
-                    subfolder = "/".join(parts[2:])
-                    controlnet = ControlNetModel.from_pretrained(
-                        repo_id,
-                        subfolder=subfolder,
-                        torch_dtype=self.dtype
-                    )
-                else:
-                    controlnet = ControlNetModel.from_pretrained(
-                        model_id,
-                        torch_dtype=self.dtype
-                    )
-            
-            controlnet = controlnet.to(device=self.device, dtype=self.dtype)
-            return controlnet
-            
-        except Exception as e:
-            raise ValueError(f"Failed to load SDXL Turbo ControlNet model '{model_id}': {e}")
-    
-    def _prepare_control_image(self, 
-                              control_image: Union[str, Image.Image, np.ndarray, torch.Tensor],
-                              preprocessor: Optional[Any] = None) -> torch.Tensor:
-        """Prepare a control image for ControlNet input"""
-        # Load image if path
-        if isinstance(control_image, str):
-            control_image = load_image(control_image)
-        
-        # Check if we can use tensor processing
-        if preprocessor is not None and isinstance(control_image, torch.Tensor) and hasattr(preprocessor, 'process_tensor'):
-            try:
-                processed_tensor = preprocessor.process_tensor(control_image)
-                if processed_tensor.dim() == 3:
-                    processed_tensor = processed_tensor.unsqueeze(0)
-                processed_tensor = processed_tensor.to(device=self.device, dtype=self.dtype)
-                
-                # SDXL: Ensure control image matches pipeline resolution
-                target_size = (self.stream.height, self.stream.width)  # (H, W)
-                current_size = processed_tensor.shape[-2:]  # (H, W)
-                if current_size != target_size:
-                    import torch.nn.functional as F
-                    processed_tensor = F.interpolate(
-                        processed_tensor,
-                        size=target_size,
-                        mode='bilinear',
-                        align_corners=False
-                    )
-                    print(f"SDXL ControlNet: Resized control image from {current_size} to {target_size}")
-                
-                return processed_tensor
-            except Exception as e:
-                print(f"⚠️  Tensor processing failed for {type(preprocessor).__name__}: {e}")
-        
-        # Direct tensor path for passthrough
-        if isinstance(control_image, torch.Tensor) and preprocessor is None:
-            target_size = (self.stream.width, self.stream.height)
-            
-            if control_image.dim() == 4:
-                control_image = control_image[0]
-            if control_image.dim() == 3 and control_image.shape[0] not in [1, 3]:
-                control_image = control_image.permute(2, 0, 1)
-            
-            current_size = control_image.shape[-2:]
-            if current_size != target_size:
-                import torch.nn.functional as F
-                if control_image.dim() == 3:
-                    control_image = control_image.unsqueeze(0)
-                
-                control_image = F.interpolate(
-                    control_image,
-                    size=target_size,
-                    mode='bilinear',
-                    align_corners=False
-                )
-                
-                if control_image.shape[0] == 1:
-                    control_image = control_image.squeeze(0)
-            
-            if control_image.dim() == 3:
-                control_image = control_image.unsqueeze(0)
-            
-            control_image = control_image.to(device=self.device, dtype=self.dtype)
-            return control_image
-        
-        # Apply preprocessor if available
-        if preprocessor is not None:
-            control_image = preprocessor.process(control_image)
-        
-        # Fast path for PIL Images
-        if isinstance(control_image, Image.Image):
-            target_size = (self.stream.width, self.stream.height)
-            if control_image.size != target_size:
-                control_image = control_image.resize(target_size, Image.LANCZOS)
-                print(f"SDXL ControlNet: Resized PIL control image to {target_size}")
-            
-            control_tensor = self._cached_transform(control_image).unsqueeze(0)
-            control_tensor = control_tensor.to(device=self.device, dtype=self.dtype)
-            return control_tensor
-        
-        # Handle other types
-        if isinstance(control_image, np.ndarray):
-            if control_image.max() <= 1.0:
-                control_image = (control_image * 255).astype(np.uint8)
-            control_image = Image.fromarray(control_image)
-        elif isinstance(control_image, torch.Tensor):
-            if control_image.dim() == 4:
-                control_image = control_image[0]
-            if control_image.dim() == 3 and control_image.shape[0] in [1, 3]:
-                control_image = control_image.permute(1, 2, 0)
-            
-            if control_image.is_cuda:
-                control_image = control_image.cpu()
-            control_image = control_image.numpy()
-            if control_image.max() <= 1.0:
-                control_image = (control_image * 255).astype(np.uint8)
-            control_image = Image.fromarray(control_image.astype(np.uint8))
-        
-        # Ensure final control image matches SDXL resolution
-        final_result = self._prepare_control_image(control_image, None)
-        
-        # Final safety check for SDXL dimensions
+        Returns:
+            Final control image tensor
+        """
+        # SDXL: Ensure control image matches pipeline resolution
         target_size = (self.stream.height, self.stream.width)  # (H, W)
-        if final_result.shape[-2:] != target_size:
+        current_size = control_tensor.shape[-2:]  # (H, W)
+        if current_size != target_size:
             import torch.nn.functional as F
-            final_result = F.interpolate(
-                final_result,
+            control_tensor = F.interpolate(
+                control_tensor,
                 size=target_size,
                 mode='bilinear',
                 align_corners=False
             )
-            print(f"SDXL ControlNet: Final resize from {final_result.shape[-2:]} to {target_size}")
+            print(f"SDXL ControlNet: Resized control image from {current_size} to {target_size}")
         
-        return final_result
-    
-    def _get_controlnet_conditioning(self, 
-                                   x_t_latent: torch.Tensor,
-                                   timestep: torch.Tensor,
-                                   encoder_hidden_states: torch.Tensor,
-                                   added_cond_kwargs: Optional[Dict[str, torch.Tensor]] = None) -> Tuple[Optional[List[torch.Tensor]], Optional[torch.Tensor]]:
-        """Get combined conditioning from all active ControlNets"""
-        if not self.controlnets:
-            return None, None
+        return control_tensor
+
+    def _get_conditioning_context(self, x_t_latent: torch.Tensor, t_list: torch.Tensor) -> Dict[str, Any]:
+        """
+        Get SDXL-specific conditioning context
         
-        active_indices = [
-            i for i, (controlnet, control_image, scale) in enumerate(
-                zip(self.controlnets, self.controlnet_images, self.controlnet_scales)
-            ) if controlnet is not None and control_image is not None and scale > 0
-        ]
-        
-        if not active_indices:
-            return None, None
-        
-        down_block_res_samples = None
-        mid_block_res_sample = None
-        
-        for i in active_indices:
-            controlnet = self.controlnets[i]
-            control_image = self.controlnet_images[i]
-            scale = self.controlnet_scales[i]
+        Args:
+            x_t_latent: Latent input
+            t_list: Timestep list
             
-            try:
-                # SDXL ControlNet call with added_cond_kwargs for time embeddings
-                controlnet_kwargs = {
-                    'sample': x_t_latent,
-                    'timestep': timestep,
-                    'encoder_hidden_states': encoder_hidden_states,
-                    'controlnet_cond': control_image,
-                    'conditioning_scale': scale,
-                    'return_dict': False,
-                }
-                
-                # Add SDXL-specific conditioning through added_cond_kwargs parameter
-                if added_cond_kwargs and ('text_embeds' in added_cond_kwargs or 'time_ids' in added_cond_kwargs):
-                    controlnet_kwargs['added_cond_kwargs'] = added_cond_kwargs
-                
-                down_samples, mid_sample = controlnet(**controlnet_kwargs)
-                
-                if down_block_res_samples is None:
-                    down_block_res_samples = down_samples
-                    mid_block_res_sample = mid_sample
-                else:
-                    for j in range(len(down_block_res_samples)):
-                        down_block_res_samples[j] += down_samples[j]
-                    mid_block_res_sample += mid_sample
-                    
-            except Exception as e:
-                print(f"Warning: SDXL Turbo ControlNet {i} failed: {e}")
-                continue
+        Returns:
+            Dictionary of SDXL conditioning context
+        """
+        # Always initialize added_cond_kwargs for SDXL compatibility
+        added_cond_kwargs = {}
         
-        return down_block_res_samples, mid_block_res_sample
-    
-    def _patch_stream_diffusion(self) -> None:
-        """Patch StreamDiffusion's unet_step method to include ControlNet conditioning"""
-        if self._is_patched:
-            return
-        
-        self._original_unet_step = self.stream.unet_step
-        
-        def patched_unet_step(x_t_latent, t_list, idx=None):
-            # Handle CFG expansion
+        # Try to get SDXL-specific conditioning from the StreamDiffusion instance
+        if hasattr(self.stream, 'add_text_embeds') and hasattr(self.stream, 'add_time_ids'):
+            # Handle text embeds expansion for CFG
             if self.stream.guidance_scale > 1.0 and (self.stream.cfg_type == "initialize"):
-                x_t_latent_plus_uc = torch.concat([x_t_latent[0:1], x_t_latent], dim=0)
-                t_list_expanded = torch.concat([t_list[0:1], t_list], dim=0)
+                add_text_embeds = torch.concat([self.stream.add_text_embeds[0:1], self.stream.add_text_embeds], dim=0)
+                add_time_ids = torch.concat([self.stream.add_time_ids[0:1], self.stream.add_time_ids], dim=0)
             elif self.stream.guidance_scale > 1.0 and (self.stream.cfg_type == "full"):
-                x_t_latent_plus_uc = torch.concat([x_t_latent, x_t_latent], dim=0)
-                t_list_expanded = torch.concat([t_list, t_list], dim=0)
+                add_text_embeds = torch.concat([self.stream.add_text_embeds, self.stream.add_text_embeds], dim=0)
+                add_time_ids = torch.concat([self.stream.add_time_ids, self.stream.add_time_ids], dim=0)
             else:
-                x_t_latent_plus_uc = x_t_latent
-                t_list_expanded = t_list
+                add_text_embeds = self.stream.add_text_embeds
+                add_time_ids = self.stream.add_time_ids
             
-            # Always initialize added_cond_kwargs for SDXL compatibility
-            added_cond_kwargs = {}
-            
-            # Try to get SDXL-specific conditioning from the StreamDiffusion instance
-            if hasattr(self.stream, 'add_text_embeds') and hasattr(self.stream, 'add_time_ids'):
-                # Handle text embeds expansion for CFG
-                if self.stream.guidance_scale > 1.0 and (self.stream.cfg_type == "initialize"):
-                    add_text_embeds = torch.concat([self.stream.add_text_embeds[0:1], self.stream.add_text_embeds], dim=0)
-                    add_time_ids = torch.concat([self.stream.add_time_ids[0:1], self.stream.add_time_ids], dim=0)
-                elif self.stream.guidance_scale > 1.0 and (self.stream.cfg_type == "full"):
-                    add_text_embeds = torch.concat([self.stream.add_text_embeds, self.stream.add_text_embeds], dim=0)
-                    add_time_ids = torch.concat([self.stream.add_time_ids, self.stream.add_time_ids], dim=0)
-                else:
-                    add_text_embeds = self.stream.add_text_embeds
-                    add_time_ids = self.stream.add_time_ids
-                
-                added_cond_kwargs = {
-                    'text_embeds': add_text_embeds,
-                    'time_ids': add_time_ids
-                }
-            elif hasattr(self.stream.pipe, 'text_encoder_2'):
-                # SDXL pipeline detected but embeddings not extracted yet
-                # Create default SDXL conditioning
-                batch_size = x_t_latent_plus_uc.shape[0]
-                
-                # Create default pooled embeddings (will be overridden by proper prompt preparation)
-                device = self.stream.device
-                dtype = self.stream.dtype
-                
-                # Default time_ids for SDXL (original_size, crops_coords_top_left, target_size)
-                time_ids = torch.tensor([
-                    [self.stream.height, self.stream.width, 0, 0, self.stream.height, self.stream.width]
-                ], dtype=dtype, device=device)
-                time_ids = time_ids.repeat(batch_size, 1)
-                
-                # Use dummy text embeds if not available
-                text_embeds = torch.zeros((batch_size, 1280), dtype=dtype, device=device)
-                
-                added_cond_kwargs = {
-                    'text_embeds': text_embeds,
-                    'time_ids': time_ids
-                }
-            
-            # Get ControlNet conditioning
-            down_block_res_samples, mid_block_res_sample = self._get_controlnet_conditioning(
-                x_t_latent_plus_uc, t_list_expanded, self.stream.prompt_embeds, added_cond_kwargs
-            )
-            
-            # Prepare UNet kwargs
-            unet_kwargs = {
-                'sample': x_t_latent_plus_uc,
-                'timestep': t_list_expanded,
-                'encoder_hidden_states': self.stream.prompt_embeds,
-                'return_dict': False,
+            added_cond_kwargs = {
+                'text_embeds': add_text_embeds,
+                'time_ids': add_time_ids
             }
+        elif hasattr(self.stream.pipe, 'text_encoder_2'):
+            # SDXL pipeline detected but embeddings not extracted yet
+            # Create default SDXL conditioning
+            batch_size = x_t_latent.shape[0]
             
-            # Add ControlNet conditioning
-            if down_block_res_samples is not None:
-                unet_kwargs['down_block_additional_residuals'] = down_block_res_samples
-            if mid_block_res_sample is not None:
-                unet_kwargs['mid_block_additional_residual'] = mid_block_res_sample
+            # Create default pooled embeddings (will be overridden by proper prompt preparation)
+            device = self.stream.device
+            dtype = self.stream.dtype
             
-            # Add SDXL-specific conditioning through added_cond_kwargs parameter
-            if added_cond_kwargs:
-                unet_kwargs['added_cond_kwargs'] = added_cond_kwargs
+            # Default time_ids for SDXL (original_size, crops_coords_top_left, target_size)
+            time_ids = torch.tensor([
+                [self.stream.height, self.stream.width, 0, 0, self.stream.height, self.stream.width]
+            ], dtype=dtype, device=device)
+            time_ids = time_ids.repeat(batch_size, 1)
             
-            # Call UNet
-            model_pred = self.stream.unet(**unet_kwargs)[0]
+            # Use dummy text embeds if not available
+            text_embeds = torch.zeros((batch_size, 1280), dtype=dtype, device=device)
             
-            # Continue with original CFG logic
-            if self.stream.guidance_scale > 1.0 and (self.stream.cfg_type == "initialize"):
-                noise_pred_text = model_pred[1:]
-                self.stream.stock_noise = torch.concat(
-                    [model_pred[0:1], self.stream.stock_noise[1:]], dim=0
-                )
-            elif self.stream.guidance_scale > 1.0 and (self.stream.cfg_type == "full"):
-                noise_pred_uncond, noise_pred_text = model_pred.chunk(2)
-            else:
-                noise_pred_text = model_pred
-            
-            if self.stream.guidance_scale > 1.0 and (
-                self.stream.cfg_type == "self" or self.stream.cfg_type == "initialize"
-            ):
-                noise_pred_uncond = self.stream.stock_noise * self.stream.delta
-            
-            if self.stream.guidance_scale > 1.0 and self.stream.cfg_type != "none":
-                model_pred = noise_pred_uncond + self.stream.guidance_scale * (
-                    noise_pred_text - noise_pred_uncond
-                )
-            else:
-                model_pred = noise_pred_text
-            
-            # Compute the previous noisy sample x_t -> x_t-1
-            if self.stream.use_denoising_batch:
-                denoised_batch = self.stream.scheduler_step_batch(model_pred, x_t_latent, idx)
-                if self.stream.cfg_type == "self" or self.stream.cfg_type == "initialize":
-                    scaled_noise = self.stream.beta_prod_t_sqrt * self.stream.stock_noise
-                    delta_x = self.stream.scheduler_step_batch(model_pred, scaled_noise, idx)
-                    alpha_next = torch.concat(
-                        [
-                            self.stream.alpha_prod_t_sqrt[1:],
-                            torch.ones_like(self.stream.alpha_prod_t_sqrt[0:1]),
-                        ],
-                        dim=0,
-                    )
-                    delta_x = alpha_next * delta_x
-                    beta_next = torch.concat(
-                        [
-                            self.stream.beta_prod_t_sqrt[1:],
-                            torch.ones_like(self.stream.beta_prod_t_sqrt[0:1]),
-                        ],
-                        dim=0,
-                    )
-                    delta_x = delta_x / beta_next
-                    init_noise = torch.concat(
-                        [self.stream.init_noise[1:], self.stream.init_noise[0:1]], dim=0
-                    )
-                    self.stream.stock_noise = init_noise + delta_x
-            else:
-                denoised_batch = self.stream.scheduler_step_batch(model_pred, x_t_latent, idx)
-            
-            return denoised_batch, model_pred
+            added_cond_kwargs = {
+                'text_embeds': text_embeds,
+                'time_ids': time_ids
+            }
         
-        self.stream.unet_step = patched_unet_step
-        self._is_patched = True
-        print("Patched StreamDiffusion with SDXL Turbo ControlNet support")
-    
-    def _unpatch_stream_diffusion(self) -> None:
-        """Restore original StreamDiffusion unet_step method"""
-        if self._is_patched and self._original_unet_step is not None:
-            self.stream.unet_step = self._original_unet_step
-            self._is_patched = False
-            print("Unpatched StreamDiffusion")
-    
-    def _convert_to_tensor_early(self, control_image: Union[str, Image.Image, np.ndarray, torch.Tensor]) -> torch.Tensor:
-        """Convert input to tensor as early as possible"""
-        if isinstance(control_image, str):
-            control_image = load_image(control_image)
+        return added_cond_kwargs
+
+    def _get_additional_controlnet_kwargs(self, **kwargs) -> Dict[str, Any]:
+        """
+        Get SDXL-specific additional kwargs for ControlNet calls
         
-        if isinstance(control_image, Image.Image):
-            import torchvision.transforms as transforms
-            to_tensor = transforms.ToTensor()
-            tensor = to_tensor(control_image)
-            return tensor.to(device=self.device, dtype=self.dtype)
+        Returns:
+            Dictionary of additional kwargs
+        """
+        # Add SDXL-specific conditioning through added_cond_kwargs parameter
+        if 'text_embeds' in kwargs or 'time_ids' in kwargs:
+            return {'added_cond_kwargs': kwargs}
+        return {}
+
+    def _get_additional_unet_kwargs(self, **kwargs) -> Dict[str, Any]:
+        """
+        Get SDXL-specific additional kwargs for UNet calls
         
-        elif isinstance(control_image, np.ndarray):
-            if control_image.max() <= 1.0:
-                tensor = torch.from_numpy(control_image).float()
-            else:
-                tensor = torch.from_numpy(control_image).float() / 255.0
-            
-            if len(tensor.shape) == 3 and tensor.shape[-1] in [1, 3]:
-                tensor = tensor.permute(2, 0, 1)
-            elif len(tensor.shape) == 2:
-                tensor = tensor.unsqueeze(0)
-                
-            return tensor.to(device=self.device, dtype=self.dtype)
-        
-        elif isinstance(control_image, torch.Tensor):
-            return control_image.to(device=self.device, dtype=self.dtype)
-        
-        else:
-            raise ValueError(f"Unsupported control image type: {type(control_image)}")
-    
-    def _tensor_to_pil_fallback(self, tensor: torch.Tensor) -> Image.Image:
-        """Fallback method to convert tensor to PIL"""
-        if tensor.dim() == 4:
-            tensor = tensor[0]
-        
-        if tensor.dim() == 3 and tensor.shape[0] in [1, 3]:
-            tensor = tensor.permute(1, 2, 0)
-        
-        if tensor.is_cuda:
-            tensor = tensor.cpu()
-        
-        if tensor.max() <= 1.0:
-            tensor = (tensor * 255).clamp(0, 255).to(torch.uint8)
-        else:
-            tensor = tensor.clamp(0, 255).to(torch.uint8)
-        
-        array = tensor.numpy()
-        
-        if array.shape[-1] == 3:
-            return Image.fromarray(array, 'RGB')
-        elif array.shape[-1] == 1:
-            return Image.fromarray(array.squeeze(-1), 'L')
-        else:
-            return Image.fromarray(array)
-    
+        Returns:
+            Dictionary of additional kwargs
+        """
+        # Add SDXL-specific conditioning through added_cond_kwargs parameter
+        if kwargs:
+            return {'added_cond_kwargs': kwargs}
+        return {}
+
     def _setup_sdxl_embeddings(self) -> None:
         """Extract and store SDXL-specific embeddings if using SDXL pipeline"""
         if not hasattr(self.stream.pipe, 'text_encoder_2'):
@@ -710,7 +218,7 @@ class SDXLTurboControlNetPipeline:
                 [self.stream.height, self.stream.width, 0, 0, self.stream.height, self.stream.width]
             ], dtype=self.stream.dtype, device=self.stream.device)
             self.stream.add_time_ids = add_time_ids.repeat(batch_size, 1)
-    
+
     def __call__(self, 
                  image: Union[str, Image.Image, np.ndarray, torch.Tensor] = None,
                  strength: float = None,
@@ -747,10 +255,6 @@ class SDXLTurboControlNetPipeline:
         from ..image_utils import postprocess_image
         output_image = postprocess_image(x_output, output_type="pil")[0]
         return output_image
-    
-    def __getattr__(self, name):
-        """Forward attribute access to the underlying StreamDiffusion instance"""
-        return getattr(self.stream, name)
 
 
 def create_sdxlturbo_controlnet_pipeline(config: StreamDiffusionControlNetConfig) -> SDXLTurboControlNetPipeline:
